Split directories
