+++ apps/image-generation/src/domain/entities/generation-result.ts
