+++ apps/image-generation/src/domain/services/llm.ts
