+++ apps/image-generation/src/domain/services/stale.ts
