+++ apps/image-generation/src/domain/services/text-to-image.ts
