+++ apps/image-generation/src/infra/repos/image-repo.ts
@@ -1,7 +1,7 @@
 import { getContext } from 'hono/context-storage'
 
-import { TrpFrogImageGenerationMetadataSchema } from '@/domain/generation-result'
-import { AITrpFrogImageRepo } from '@/domain/image-repo'
+import { TrpFrogImageGenerationMetadataSchema } from '@/domain/entities/generation-result'
+import { AITrpFrogImageRepo } from '@/domain/repos/image-repo'
 import { Env } from '@/env'
 
 const CURRENT_IMAGE_KEY = 'current-image'
