+++ apps/image-generation/src/infra/services/llm.ts
