+++ apps/image-generation/src/infra/services/text-to-image.ts
